API-417: Move all tests in Common namespace
