+++ phpunit.xml.dist
@@ -15,9 +15,4 @@
         <directory suffix="Integration.php">tests/</directory>
     </testsuite>
 
-    <testsuite name="PHP_Client_Integration_Test_2_0">
-        <directory suffix="Integration.php">tests/Common/Api/*</directory>
-        <directory suffix="Integration.php">tests/v2_0/Api/*</directory>
-    </testsuite>
-
 </phpunit>
